refactor: extracts tabed pages into components
